+++ testFiles/mixinTarget.js
@@ -1,0 +1,11 @@
+define(['mixinTargetBase'], function (Base) {
+  return Base.extend({
+    defaults: {
+      category: null
+    },
+
+    check () {
+      return true
+    }
+  })
+})